Move preprocessing functions used for models trained on ImageNet to the imagerecognition package
