+++ api/src/jvmMain/kotlin/org/jetbrains/kotlinx/dl/api/inference/imagerecognition/InputType.kt
@@ -3,7 +3,7 @@
  * Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE.txt file.
  */
 
-package org.jetbrains.kotlinx.dl.api.inference.keras.loaders
+package org.jetbrains.kotlinx.dl.api.inference.imagerecognition
 
 /**
  * Different Neural Networks were trained on ImageNet dataset with different image preprocessing.
